fix wind and temp

+++ service-weather/src/main/java/com/gogoaren/indarra/serviceweather/fetch/Weather.java
@@ -3,19 +3,14 @@ package com.gogoaren.indarra.serviceweather.fetch;
 import lombok.Builder;
 import lombok.Data;
 
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import java.io.Serializable;
-import java.time.Instant;
-import java.util.UUID;
+import java.math.BigDecimal;
 
 @Data
 @Builder
 public class Weather {
 
     private String city;
-    private double temperature;
+    private BigDecimal temperature;
     private double humidity;
     private double wind;
